BitmapFileMaterial : Now releases BitmapDatas that are no longer being used, and dispatches LOAD_COMPLETE events properly, even for files that are already cached.
DAE : Now has a public property forceCoordSet which you can change to 0 if there are multiple UV definitions in the DAE and your materials aren't mapping on correctly. I've also changed all private properties and methods to protected. I have also enabled a flag for asynchronous parsing of the XML on the load function. It doesn't currently seem to work. Ongoing investigations. :-) 

DisplayObjectContainer3D : now checks to see if an added child is already in the scene somewhere, and logs an error with PaperLogger.